server/requirements.txt
flask==2.2.5
gunicorn==21.2.0
flask_sqlalchemy==3.1.1
python-dotenv===1.0.1
psycopg2-binary==2.9.9
jsonify===0.5
flask-cors===4.0.1

